+++ CHANGELOG.rst
@@ -26,6 +26,9 @@ new version (on deck)
   parameter, allowing dataframe merging, if the result of the text processing is a
   dataframe.@samukweku
 - [ENH] Add ``aggfunc`` parameter to pivot_wider. @samukweku
+- [ENH] Modified the `check` function in utils to verify if a value is a callable. @samukweku
+- [ENH] Add a base `_select_column` function, using ``functools.singledispatch``,
+  to allow for flexible columns selection. @samukweku
 
 
 
@@ -42,6 +45,7 @@ v0.20.10
   tidyr's ``pivot_longer`` function. Also added an example notebook. @samukweku
 - [ENH] Fixed code to returns error if `fill_value` is not a dictionary. @samukweku
 - [INF] Welcome bot (.github/config.yml) for new users added. Issue #739. @samukweku
+